requirements.txt
streamlit==1.10.0
google-generativeai==0.1.0